chore: refactor imports and remove MainSection maxW
