+++ packages/apps/spaces/app/organizations/[id]/components/Timeline/events/email/EmailTimelineItem.tsx
@@ -1,7 +1,7 @@
 'use client';
 import React, { FC } from 'react';
 import Image from 'next/image';
-import { Card, CardBody, CardFooter } from '@ui/layout/Card';
+import { Card, CardBody, CardFooter } from '@ui/presentation/Card';
 import { Text } from '@ui/typography/Text';
 import { VStack } from '@chakra-ui/react';
 import { convert } from 'html-to-text';
