--- packages/apps/spaces/ui/layout/Card/index.ts
@@ -1,7 +1,0 @@
-export type {
-  CardProps,
-  CardHeaderProps,
-  CardBodyProps,
-  CardFooterProps,
-} from '@chakra-ui/react';
-export { Card, CardHeader, CardBody, CardFooter } from '@chakra-ui/react';
