+++ packages/apps/spaces/ui/presentation/Card/index.ts
@@ -1,0 +1,7 @@
+export type {
+  CardProps,
+  CardBodyProps,
+  CardFooterProps,
+  CardHeaderProps,
+} from '@chakra-ui/react';
+export { Card, CardBody, CardFooter, CardHeader } from '@chakra-ui/react';
